abstract ConfirmButton into a separate element that ConfirmDeleteButton uses
